+++ kernel/core/maths/filtering.hpp
@@ -51,7 +51,7 @@ namespace NumeRe
         size_t col;
         double alpha;
 
-        FilterSettings(FilterType _type = FILTER_NONE, size_t _row = 1u, size_t _col = 1u, double _alpha = NAN) : type(_type), row(std::max(_row, 1u)), col(std::max(_col, 1u)), alpha(_alpha)
+        FilterSettings(FilterType _type = FILTER_NONE, size_t _row = 1u, size_t _col = 1u, double _alpha = NAN) : type(_type), row(std::max(_row, (size_t)1u)), col(std::max(_col, (size_t)1u)), alpha(_alpha)
         {
             //
         }
@@ -135,7 +135,7 @@ namespace NumeRe
             {
                 // This expression avoids that someone tries to
                 // create a filter with a zero dimension.
-                m_windowSize = std::make_pair(std::max(row, 1u), std::max(col, 1u));
+                m_windowSize = std::make_pair(std::max(row, (size_t)1u), std::max(col, (size_t)1u));
             }
 
             /////////////////////////////////////////////////
@@ -276,9 +276,9 @@ namespace NumeRe
                 {
                     for (size_t j = 0; j < m_windowSize.second; j++)
                     {
-                        if (sqrt(pow2(i/((double)std::max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max(1u, m_windowSize.second-1))-mean_col)) <= 0.5)
+                        if (sqrt(pow2(i/((double)std::max((size_t)1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max((size_t)1u, m_windowSize.second-1))-mean_col)) <= 0.5)
                         {
-                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/((double)std::max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max(1u, m_windowSize.second-1))-mean_col)) - 0.5);
+                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/((double)std::max((size_t)1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max((size_t)1u, m_windowSize.second-1))-mean_col)) - 0.5);
                             sum += m_filterKernel[i][j];
                         }
                     }
@@ -868,9 +868,9 @@ namespace NumeRe
                 {
                     for (size_t j = 0; j < m_windowSize.second; j++)
                     {
-                        if (sqrt(pow2(i/((double)std::max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max(1u, m_windowSize.second-1))-mean_col)) <= 0.5)
+                        if (sqrt(pow2(i/((double)std::max((size_t)1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max((size_t)1u, m_windowSize.second-1))-mean_col)) <= 0.5)
                         {
-                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/((double)std::max(1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max(1u, m_windowSize.second-1))-mean_col)) - 0.5);
+                            m_filterKernel[i][j] = fabs(sqrt(pow2(i/((double)std::max((size_t)1u, m_windowSize.first-1))-mean_row) + pow2(j/((double)std::max((size_t)1u, m_windowSize.second-1))-mean_col)) - 0.5);
                         }
                     }
                 }
